pyproject.toml
[tool.poetry]
name = "pre-delinquency-engine"
version = "0.1.0"
description = "Pre-delinquency intervention system using behavioral ML"
authors = ["Your Team <team@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"

# Data & Database
pandas = "^2.2.0"
numpy = "^1.26.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.0"
faker = "^24.0.0"

# ML & Feature Engineering
xgboost = "^2.0.0"
lightgbm = "^4.3.0"
scikit-learn = "^1.4.0"
shap = "^0.45.0"
mlflow = "^2.11.0"
joblib = "^1.3.0"

# API & Backend
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
python-multipart = "^0.0.9"
websockets = "^12.0"

# Dashboard & Visualization
streamlit = "^1.32.0"
plotly = "^5.19.0"
altair = "^5.2.0"

# Development & Testing
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
ipykernel = "^6.29.0"
jupyter = "^1.0.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
